[Chapter1] refactor: createStatementData, statement 함수를 두개의 파일로 분리
