+++ src/Chapter1/statement.js
@@ -1,77 +1,7 @@
+import createStatementData from './createStatementData';
+
 function statement(invoice, plays) {
   return renderPlainText(createStatementData(invoice, plays));
-
-  function createStatementData(invoice, plays) {
-    const statementData = {};
-
-    statementData.customer = invoice.customer;
-    statementData.performances = invoice.performances.map(enrichPerformance);
-    statementData.totalAmount = totalAmount(statementData);
-    statementData.totalVolumeCredits = totalVolumeCredits(statementData);
-
-    return statementData;
-
-    function enrichPerformance(aPerformance) {
-      const result = Object.assign({}, aPerformance); //얕은 복사 수행
-
-      result.play = playFor(result); //중간 데이터에 연극 정보를 저장
-      result.amount = amountFor(result);
-      result.volumeCredits = volumeCreditsFor(result);
-
-      return result;
-    }
-
-    function playFor(aPerformance) {
-      return plays[aPerformance.playID];
-    }
-
-    function amountFor(aPerformance) {
-      let result = 0;
-
-      switch (aPerformance.play.type) {
-        //비극
-        case 'tragedy':
-          result = 40000;
-          if (aPerformance.audience > 30) {
-            result += 1000 * (aPerformance.audience - 30);
-          }
-          break;
-
-        //희극
-        case 'comedy':
-          result = 30000;
-          if (aPerformance.audience > 20) {
-            result += 10000 + 500 * (aPerformance.audience - 20);
-          }
-          result += 300 * aPerformance.audience;
-          break;
-
-        default:
-          throw new Error(`알 수 없는 장르: ${aPerformance.play.type}`);
-      }
-
-      return result;
-    }
-
-    function volumeCreditsFor(aPerformance) {
-      let volumeCredits = 0;
-      volumeCredits += Math.max(aPerformance.audience - 30, 0);
-
-      if ('comedy' === aPerformance.play.type) {
-        volumeCredits += Math.floor(aPerformance.audience / 5);
-      }
-
-      return volumeCredits;
-    }
-
-    function totalAmount(data) {
-      return data.performances.reduce((total, p) => total + p.amount, 0);
-    }
-
-    function totalVolumeCredits(data) {
-      return data.performances.reduce((total, p) => total + p.volumeCredits, 0);
-    }
-  }
 }
 
 function renderPlainText(data, plays) {
